Calculate standardized module scores for feature expression programs in single cells.

Calculates the standardized expression level of each program on single-cell level, by subtracting by aggregated expression of control feature sets (like in `AddModuleScore`) and then scaling the difference by the pooled variance of the gene set of interest and control features. Like `AddModuleScore`, all analyzed features are binned based on averaged expression and control features are randomly selected from each bin.

AddSModuleScore(
  object,
  features,
  pool = NULL,
  nbin = NULL,
  ctrl = 100,
  nworkers = 1,
  scale = T,
  k = FALSE,
  assay = NULL,
  name = "Cluster",
  seed = 1,
  search = FALSE,
  ...
)

Arguments

object

Seurat Object

features

A list of vectors of features for expression programs; each entry should be a vector of feature names

pool

List of features to check expression levels against, defaults to `rownames(x = object)`

nbin

Number of bins of aggregate expression levels for all analyzed features

ctrl

Number of control features selected from the same bin per analyzed feature

nworkers

Number of workers for parallel implementation. Default is 1.

scale

scale score by pooled variance. Default is T.

k

Use feature clusters returned from DoKMeans

assay

Name of assay to use

name

Name for the expression programs; will append a number to the end for each entry in features (eg. if `features` has three programs, the results will be stored as name1, name2, name3, respectively)

seed

Set a random seed. If NULL, seed is not set.

search

Search for symbol synonyms for features in features that don't match features in object? Searches the HGNC's gene names database; see UpdateSymbolList for more details

...

Extra parameters passed to UpdateSymbolList

Value

Returns seurat object with standardized module scores added to object meta data; each module is stored as name# for each module program present in `features`

See also

Author

Nicholas Mikolajewicz