Gene expression markers for all identity classes.

Finds markers (differentially expressed genes) for each of the identity classes in a dataset, with parallel implementation. Produces identical output as Seurat::FindAllMarkers,

FindAllMarkers.Parallel(
  object,
  n.work = 1,
  assay = DefaultAssay(object),
  slot = "data",
  only.pos = F,
  min.pct = 0,
  test.use = "MAST",
  logfc.threshold = 0.15,
  max.cells.per.ident = 200,
  return.thresh = 1,
  verbose = F,
  ...
)

Arguments

object
Seurat object

n.work
Number of workers. If 1 (default), runs Seurat::FindAllMarkers(). If >1, parallelized implementation is used.

assay
Assay to use in differential expression testing. Default is DefaultAssay(object)

slot
Slot to pull data from; note that if test.use is "negbinom", "poisson", or "DESeq2", slot will be set to "counts". Default is "data".

only.pos
Only return positive markers (FALSE by default)

min.pct
only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0

test.use
Test to use. See Seurat::FindAllMarkers() or Seurat::FindMarkers() for list of all options.

logfc.threshold
Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25 Increasing logfc.threshold speeds up the function, but can miss weaker signals.

max.cells.per.ident
Down sample each identity class to a max number. Default is 200.

return.thresh
Only return markers that have a p-value < return.thresh, or a power > return.thresh (if the test is ROC).

verbose
Print a progress bar once expression testing begins. Default is F.

...
additional arguments passed to Seurat::FindAllMarkers() (if n.work = 1) or Seurat::FindMarkers() (if n.work >1)

See also

Author

Nicholas Mikolajewicz

Examples

# Find markers for all clusters
all.markers <- FindAllMarkers.Parallel(object = so.query, n.work = 4)

#> Loading required package: BiocParallel

#> Error in unique(object@active.ident): object 'so.query' not found