Load parameter specifications for Module 1

Load parameter specification list for Module 1 (preprocessing and QC). If not specified, default settings are applied.

analysisParameters(
  which.data,
  which.strata = NULL,
  organism.filter.flag,
  organism.include,
  save.flag = T,
  save.filename,
  save.directory = "Preprocessed_Datasets/",
  set_names = NULL,
  data.imputed.flag = F,
  vars2regress = "percent.mt",
  subsample_factor = 1,
  plt.log.flag = TRUE,
  gene.upperlimit = 9000,
  gene.lowerlimit = 200,
  mt.upperlimit = 60,
  unmatched.rate.filter.flag = T,
  do.additional.reduction = F,
  cluster.resolution = 0.4,
  print.inline = FALSE
)

Arguments

which.data
Character. Specify input data (input file specification are stored in separate data.frame. See Module 1 for details)

which.strata
Character. Specify how input data are subset. Leave unspecified to keep all data.

organism.filter.flag
Logical specifying whether inputs should be filtered by species. Recommended if multiple species are present in input.

organism.include
Character. Which species to include. One of:

"Hs" - Human

"Mm" - Mouse

"Hs, Mm" - both species included

save.flag
Logical specifying whether preprocessed data (seurat object) are saved as RData file. Default is TRUE

save.filename
Character. Output filename (must include .Rdata suffix)

save.directory
Character. output directory (Default is "Preprocessed Datasets/")

set_names
Optional character. Label for dataset.

data.imputed.flag
Logical specifying wehther data were imputed. Default is F.

vars2regress
Character vector specifying which parameters to regress during data scaling. Default is "percent.mt"

subsample_factor
Numeric [0,1]. Factor used to subsample data. Recommended during initial data exploration, especially if dataset is large.

plt.log.flag
Logical specifying whether QC violin plots are on log scale.

gene.upperlimit
Numeric. Upper limit for number of genes per cell. Default is 200.

gene.lowerlimit
Numeric. Lower limit for number of genes per cell. Default is 9000.

mt.upperlimit
Numeric [0,100]. Upper limit for mitochondrial percentage. Default is 60.

unmatched.rate.filter.flag
Logical flag for filtering cells unmatched rate. Adaptive threshold are used.

do.additional.reduction
Flag to generate ICA and TSNE reductions. Default is FALSE.

cluster.resolution
Resolution used for cell clustering.

print.inline
Logical specifying wheter data are printed in R notebook. Default is F - recommended if generated flexdashboard output.

Value

list