Word cloud visualization of cell-type annotations from the Miko scoring pipeline

Given outputs from the Miko scoring pipeline, the top cell-type annotations fare visualized using word clouds.

annotationCloud(
  object,
  object.group = "seurat_clusters",
  score,
  score.group,
  score.cell.type,
  score.p,
  score.fdr = NULL,
  score.coherence.fraction = NULL,
  score.frequent.flier = NULL,
  fdr.correction = T,
  p.threshold = 0.05,
  coherence.threshold = 0.8,
  show.n.terms = 15,
  verbose = T
)

Arguments

object

Seurat Object.

object.group

name of object meta data field specifying cluster membership. Default is "seurat_clusters".

score

vector of Miko scores

score.group

vector of group memberships

score.cell.type

vector of cell-type names/labels.

score.p

vector of p values.

score.fdr

vector of fdr values. Optional.

score.coherence.fraction

vector of coherence fractions. See coherentFraction(...) for details.

score.frequent.flier

vector of logicals specifying whether score belongs to frequent flier.

fdr.correction

Specify whether p-value should be corrected using Benjamini & Hochberg method. Default is T.

p.threshold

p value threshold. Default is 0.05.

coherence.threshold

Numerical [0,1] specifying minimal coherence required to qualify for visualization. Default is 0.8.

show.n.terms

Maximal number of cell-type terms shown in word cloud. Default is 15.

verbose

Logical, specify whether process is printed. Default is T.

Value

list of ggplot handles

See also

Author

Nicholas Mikolajewicz

Examples

df.score_summary <- data.frame(cluster = df.merge$cluster,
                              cell.type = df.merge$gs,
                              miko_score = signif(df.merge$miko_score, 3) ,
                              p =  signif(df.merge$p),
                              fdr =  signif(df.merge$fdr),
                              coherence_fraction =  signif(df.merge$coherence_fraction))
#> Error in data.frame(cluster = df.merge$cluster, cell.type = df.merge$gs,     miko_score = signif(df.merge$miko_score, 3), p = signif(df.merge$p),     fdr = signif(df.merge$fdr), coherence_fraction = signif(df.merge$coherence_fraction)): object 'df.merge' not found

plt.cloud <- annotationCloud(object = so.query_scored,
                             object.group = "seurat_clusters",
                             score = df.score_summary$miko_score,
                             score.group = df.score_summary$cluster,
                             score.cell.type = df.score_summary$cell.type,
                             score.p = df.score_summary$p,
                             score.fdr = df.score_summary$fdr,
                             score.coherence.fraction = df.score_summary$coherence_fraction,
                             score.frequent.flier = NULL,
                             fdr.correction = T,
                             p.threshold = 0.05,
                             coherence.threshold = 0.9,
                             show.n.terms = 15,
                             verbose = T)
#> Loading required package: ggwordcloud
#> 2022-08-03 14:41:57: Generating annotation wordclouds...
#> Error in annotationCloud(object = so.query_scored, object.group = "seurat_clusters",     score = df.score_summary$miko_score, score.group = df.score_summary$cluster,     score.cell.type = df.score_summary$cell.type, score.p = df.score_summary$p,     score.fdr = df.score_summary$fdr, score.coherence.fraction = df.score_summary$coherence_fraction,     score.frequent.flier = NULL, fdr.correction = T, p.threshold = 0.05,     coherence.threshold = 0.9, show.n.terms = 15, verbose = T): object 'df.score_summary' not found