Analysis of scale free topology for soft-threshold. Modified from getSoftThreshold.

Analysis of scale free topology for multiple soft thresholding powers. The aim is to help the user pick an appropriate soft-thresholding power for network construction. Inspired by WGCNA::pickSoftThreshold and updated from first version (scMiko::getSoftThreshold)

getSoftThreshold.v2(
  s.mat,
  power = c(seq(0.5, 5, by = 0.5), seq(6, 10)),
  network.type = "signed",
  nBreaks = 20,
  removeFirst = T,
  rescale.adjacency = F
)

Arguments

s.mat
similarity matrix

power
Numeric vector of powers to evaluate. Default is c(seq(0.5,5, by = 0.5), seq(6,10))

nBreaks
Number of bins in connectivity histograms. Default is 20.

removeFirst
Logical specifying whether the first bin should be removed from the connectivity histogram. Default is True.

rescale.adjacency
Logical indicating if s.mat should be rescaled to [0,1]

networkType
Allowed values are (unique abbreviations of) "unsigned", "signed", "signed hybrid". See WGCNA::adjacency.

Value

named list containing power estimates, r2 estimates, distribution plots, optimiation plot and results data.frame.

Examples

# determine optimal soft threshold
sft <- getSoftThreshold.v2(s.mat, power =c(seq(0.5,5, by = 0.5), seq(6,10)),
network.type = "signed", rescale.adjacency = F)

#> Error in sim2adj(s.mat, soft.power = power.cur, network.type = network.type): object 's.mat' not found

# visualize optimization plot
print(sft$optimization.plot)

#> Error in print(sft$optimization.plot): object 'sft' not found

# visualize node-linkage density plots
cowplot::plot_grid(plotlist = sft$distribution.plot, ncol = 5)

#> Error in cowplot::plot_grid(plotlist = sft$distribution.plot, ncol = 5): object 'sft' not found