Analysis of scale free topology for soft-threshold. Modified from getSoftThreshold.

Analysis of scale free topology for multiple soft thresholding powers. The aim is to help the user pick an appropriate soft-thresholding power for network construction. Inspired by WGCNA::pickSoftThreshold and updated from first version (scMiko::getSoftThreshold)

getSoftThreshold2(
  s.mat,
  power = c(seq(0.5, 5, by = 0.5), seq(6, 10)),
  network.type = "signed",
  nBreaks = 20,
  removeFirst = T,
  rescale.adjacency = F,
  n.cores = 4,
  r2.target = 0.9
)

Arguments

s.mat
similarity matrix

power
Numeric vector of powers to evaluate. Default is c(seq(0.5,5, by = 0.5), seq(6,10))

nBreaks
Number of bins in connectivity histograms. Default is 20.

removeFirst
Logical specifying whether the first bin should be removed from the connectivity histogram. Default is True.

rescale.adjacency
Logical indicating if s.mat should be rescaled to [0,1]

n.cores
Number of cores to use for parallelization. Default is 4.

r2.target
Retrieves soft power that corresponds to network topology corresponding to r2 >= r2.target. Default is 0.9.

networkType
Allowed values are (unique abbreviations of) "unsigned", "signed", "signed hybrid". See WGCNA::adjacency.

Value

named list containing power estimates, r2 estimates, distribution plots, optimiation plot and results data.frame.

See also

Examples

# determine optimal soft threshold
sft <- getSoftThreshold2(s.mat, power =c(seq(0.5,5, by = 0.5), seq(6,10)),
network.type = "signed", rescale.adjacency = F)

#> Error in (foreach(i = 1:length(powers), .packages = c("scMiko", "dplyr",     "ggplot2", "ggpmisc"))) %dopar% {    power.cur <- powers[i]    a.cur <- sim2adj(s.mat, soft.power = power.cur, network.type = network.type)    if (rescale.adjacency)         a.cur <- recaleValues(a.cur, new.min = 0, new.max = 1)    net.connectivity.df <- getConnectivity(a.cur, rownames(a.cur),         flag.top.n = 20)    rm(a.cur)    invisible({        gc()    })    k <- net.connectivity.df$wi    discretized.k <- cut(k, nBreaks)    dk <- tapply(k, discretized.k, mean)    p.dk <- as.vector(tapply(k, discretized.k, length)/length(k))    breaks1 <- seq(from = min(k), to = max(k), length = nBreaks +         1)    hist1 <- suppressWarnings(hist(k, breaks = breaks1, equidist = FALSE,         plot = FALSE, right = TRUE))    dk2 <- hist1$mids    dk <- ifelse(is.na(dk), dk2, dk)    dk = ifelse(dk == 0, dk2, dk)    p.dk <- ifelse(is.na(p.dk), 0, p.dk)    log.dk <- as.vector(log10(dk))    if (removeFirst) {        p.dk = p.dk[-1]        log.dk = log.dk[-1]    }    log.p.dk <- as.numeric(log10(p.dk + 1e-09))    lm1 <- lm(log.p.dk ~ log.dk)    r2.sf.current <- summary(lm1)[["r.squared"]] * sign(lm1[["coefficients"]][["log.dk"]])    df.sf <- data.frame(x = log.dk, y = log.p.dk)    plt.sf.current <- df.sf %>% ggplot(aes(x = x, y = y)) + geom_smooth(method = "lm",         color = "tomato", fill = "tomato") + geom_point(size = 3) +         xlab("N Links (Log)") + ylab("N Nodes (Log)") + labs(title = "Node Linkages",         subtitle = paste0("Soft Power = ", power.cur)) + theme_classic() +         stat_fit_glance(method = "lm", label.y = "bottom", method.args = list(formula = y ~             x), mapping = aes(label = sprintf("r^2~\"=\"~%.3f~~italic(P)~\"=\"~%.2g",             stat(r.squared), stat(p.value))), parse = TRUE)    sf.list <- list(r2 = r2.sf.current, df = df.sf, plt = plt.sf.current)    return(sf.list)}: could not find function "%dopar%"

# visualize optimization plot
print(sft$optimization.plot)

#> Error in print(sft$optimization.plot): object 'sft' not found

# visualize node-linkage density plots
cowplot::plot_grid(plotlist = sft$distribution.plot, ncol = 5)

#> Error in cowplot::plot_grid(plotlist = sft$distribution.plot, ncol = 5): object 'sft' not found