Infer initial trajectory through space, using pseudotime priors

Infer initial trajectory through space, using pseudotime priors. Adopted from destiny package, uses smth.gaussian function; The specific function for smoothing using the gaussian window function.

inferInitialTrajectory.v2(pt, space, w_width = 0.1)

Arguments

pt
pseudotimes. Numeric vector.

space
A numeric matrix or a data frame containing the coordinates of samples.

w.width
the length of the smoothing window, if an integer, represents number of items, else, if a value between 0 and 1, represents the proportion of the input vector

Value

the initial trajectory

See also

Examples

 # lineage-specific pseudotime
 pt_vec <- dpt_for_branch(dpt_flat, branch_id = paths_to)

#> Error in dpt_for_branch(dpt_flat, branch_id = paths_to): object 'dpt_flat' not found

# get indices
idx <- dpt_flat@branch[, 1] %in% c(root, paths_to)

#> Error in dpt_flat@branch[, 1] %in% c(root, paths_to): object 'dpt_flat' not found

# get initial trajectory path
umap.path <- inferInitialTrajectory.v2(pt_vec[idx], so.query@reductions[["umap"]]@cell.embeddings[idx, ], w_width = 0.15)

#> Error in nrow(space): object 'so.query' not found

   # get umap coordinates
df.umap.all <- data.frame(x = so.query@reductions[["umap"]]@cell.embeddings[ ,1],
                          y = so.query@reductions[["umap"]]@cell.embeddings[ ,2])

#> Error in data.frame(x = so.query@reductions[["umap"]]@cell.embeddings[,     1], y = so.query@reductions[["umap"]]@cell.embeddings[, 2]): object 'so.query' not found

df.umap.all$col <- "grey"

#> Error in df.umap.all$col <- "grey": object 'df.umap.all' not found

df.umap.sub <- df.umap.all[idx, ]

#> Error in eval(expr, envir, enclos): object 'df.umap.all' not found

df.umap.sub$pt <- pt_vec[idx]

#> Error in eval(expr, envir, enclos): object 'pt_vec' not found

# compute prinicpal curves
pc.fit <- princurve::principal_curve(as.matrix(df.umap.sub[,c("x", "y")]), start = as.matrix(umap.path[ ,c(1,2)]),
                                     thresh = 0.001, maxit = 10, stretch = 2, smoother = "smooth_spline",
                                     approx_points = 100, trace = FALSE, plot_iterations = FALSE)

#> Error in as.matrix(df.umap.sub[, c("x", "y")]): object 'df.umap.sub' not found

traj.path <- pc.fit$s[pc.fit$ord, , drop = FALSE]

#> Error in eval(expr, envir, enclos): object 'pc.fit' not found

df.tp <- data.frame(traj.path)

#> Error in data.frame(traj.path): object 'traj.path' not found