Get lineage trajectories using principal curves and compute pseudotimes

For given space (e.g., UMAP, PCA, etc.), lineage trajectories are fit using prinicpal curves, and these are then used to derive pseudotimes.

lineageTrajectory(
  space,
  start = NULL,
  group.labels = NULL,
  pseudotimes = NULL,
  thresh = 0.001,
  maxit = 10,
  stretch = 2,
  smoother = "smooth_spline",
  approx_points = 100,
  trace = FALSE,
  plot_iterations = FALSE
)

Arguments

space
A numeric matrix or a data frame containing the coordinates of samples.

start
either a previously fit principal curve, or else a matrix of points that in row order define a starting curve. If missing or NULL, then the first principal component is used. If the smoother is "periodic_lowess", then a circle is used as the start.

group.labels
Character vector (same length as number of rows in space) specifying group membership. Optional.

pseudotimes
Either logical (TRUE) specifying whether to use internally generated pseudotimes (from principal curve fitting) to color generated umap points, or, numeric vector with pseudotimes. If provided, plotted data are colored by pseudotime. Otherwise colored by group membership.

thresh
convergence threshold on shortest distances to the curve. Default is 0.001.

maxit
maximum number of iterations.

stretch
A stretch factor for the endpoints of the curve, allowing the curve to grow to avoid bunching at the end. Must be a numeric value between 0 and 2.

smoother
choice of smoother. The default is "smooth_spline", and other choices are "lowess" and "periodic_lowess". The latter allows one to fit closed curves. Beware, you may want to use iter = 0 with lowess().

approx_points
Approximate curve after smoothing to reduce computational time. If FALSE, no approximation of the curve occurs. Otherwise, approx_points must be equal to the number of points the curve gets approximated to; preferably about 100.

trace
If TRUE, the iteration information is printed

plot_iteractions
If TRUE the iterations are plotted.

Value

list of results containing principal curve fits and coordinates, pseutimes, plots

See also

Author

Nicholas Mikolajewicz

Examples