Load preprocessed data from Moffat lab sciRNA-seq3 pipeline

Load preprocessed data from Moffat lab sciRNA-seq3 pipeline. RT barcode and plate summary is stored in misc slot of resulting seurat object.

loadMoffat(
  import_set,
  subsample_factor,
  input_organisms,
  organism_include,
  dir
)

Arguments

import_set
Character vector specifying expression matrix (import_set[1]), PCR barcodes (import_set[2]) and RC barcodes (import_set[3]). Expression matrix will be imported successfully if barcodes are omitted.

subsample_factor
Numeric [0,1]. Subsampling factor

input_organisms
All species included in input files. One of:

"Hs" - Human

"Mm" - Mouse

c("Hs", "Mm") - both species included

organism_include
Species to include in downstream analysis. One of:

"Hs" - Human

"Mm" - Mouse

c("Hs", "Mm") - both species included

dir
Character. folder containing import_set files

Value

list containing Seurat Object and named gene vector.