scRNAseq integration wrapper

scRNAseq normalization and integration wrapper. Given seurat object input, data are split, normalized and integrated.

miko_integrate(
  object,
  split.by = "Barcode",
  min.cell = 50,
  k.anchor = 20,
  k.weight = 35,
  nfeatures = 3000,
  split.prenorm = F,
  assay = "SCT",
  variable.features.n = 3000,
  verbose = T,
  use.existing.sct = F,
  conserve.memory = F,
  vars.to.regress = "percent.mt"
)

Arguments

object
Seurat object

split.by
Meta data feature to split and integrate data.

min.cell
Minimum number of cells permitted per object prior to integration.

k.anchor
How many neighbors (k) to use when picking anchors.

k.weight
Number of neighbors to consider when weighting anchors.

nfeatures
Number of features to return (passed to SelectIntegrationFeatures)

split.prenorm
Split data before (TRUE) or after (FALSE) normalization.

assay
Assay to use for normalization.

variable.features.n
Use this many features as variable features after ranking by residual variance; default is 3000.

verbose
Print progress. Default is TRUE.

use.existing.sct
If TRUE, existing SCT model is used. Default is FALSE (new SCT model is fit)

conserve.memory
If set to TRUE the residual matrix for all genes is never created in full when running SCTransform; useful for large data sets, but will take longer to run; this will also set return.only.var.genes to TRUE; default is FALSE

vars.to.regress
meta features to regress out. Default is "percent.mt". Set to NULL if unspecified.

Value

Integrated seurat object

See also

Author

Nicholas Mikolajewicz

Examples