Perform differential expression analysis

Give two existing fields within a Seurat Object, data are stratified by the first, and pairwise comparisons between groups in the second are performed to identify differential gene expression. Uses Seurat's FindMarkers() function.

multiLevel.FindMarkers(so, ordered.levels, which.assay = NULL, ...)

Arguments

so
Seurat object.

ordered.levels
Vector of one or two seurat meta fields. If one provided, the second is set to "seurat_clusters".

which.assay
Seurat assay to use. If unspecified, set to DefaultAssay(so).

...
Additional arguments passed to Seurat's FindMarkers() function.

Value

list of data.frames where each data.fram contains results from an individual pairwise comparison.

Examples

# define mapping lists
mapping.list.1 <- list(NSG = "NSG", BALBc = "BALB")
mapping.list.2 <- list(Early = "Early", Mid = "Mid", Late = "Late")

# Create new field called "Mouse" using relabled entries from "Condition" - relabelling based on mapping list
so.query.2 <- mapSubsetSeurat(so.query, old.field = "Condition", new.field = "Mouse", mapping.list = mapping.list.1)

#> Error in grepl(cur.pattern, so@meta.data[[old.field]]): object 'so.query' not found

# Create new field called "Time" using relabled entries from "Group" - relabelling based on mapping list
so.query.2 <- mapSubsetSeurat(so.query.2, old.field = "Group", new.field = "Time", mapping.list = mapping.list.2)

#> Error in grepl(cur.pattern, so@meta.data[[old.field]]): object 'so.query.2' not found

# Perform multilevel differential gene expression.
# Data are stratified by time, and then pairwise comparison between "Mouse" groups are performed.
# in this example, 3 total pairwise comparisons were made: BALBc vs NSG (Early), BALBc vs NSG (Mid) and BALBc vs NSG (Late)
deg.list <- multiLevel.FindMarkers(so.query.2, ordered.levels = c("Time", "Mouse"),
                                   logfc.threshold = 0,
                                   min.pct = 0,
                                   test.use = "wilcox")

#> Error in DefaultAssay(so): could not find function "DefaultAssay"