Null distribution for standardized module scores.

Downsample number of cells in Seurat object by specified factor

nullScore(
  object,
  assay = DefaultAssay(object),
  n.replicate = 25,
  min.gs.size = 2,
  max.gs.size = 100,
  step.size = 2,
  nworkers = 1,
  subsample.n = NULL,
  nbin = 24,
  verbose = T
)

Arguments

object
Seurat Object

assay
Name of assay to use

n.replicate
Number of replicates

min.gs.size
minimum gene set size

max.gs.size
maximum gene set size

step.size
step size; increment of sequence between `min.gs.size` and `max.gs.size`.

nworkers
number of workers used for parallel implementation. Default is 1.

subsample.n
Numeric [1,ncol(object)]. Number of cells to subsample. If specified, overides subsample.factor.

nbin
Number of bins of aggregate expression levels for all analyzed features. See `AddModuleScore` for details.

verbose
Print progress. Default is TRUE.

Value

list of results that are used as input into `mikoScore`.

See also

Author

Nicholas Mikolajewicz