Quickly estimate the 'elbow' of a scree plot (PCA)

This function uses a rough algorithm to estimate a sensible 'elbow' to choose for a PCA scree plot of eigenvalues. The function looks at an initial arbitrarily 'low' level of variance and looks for the first eigenvalue lower than this. If the very first eigenvalue is actually lower than this (i.e, when the PCs are not very explanatory) then this 'low' value is iteratively halved until this is no longer the case. After starting below this arbitrary threshold the drop in variance explained by each pair of consecutive PCs is standardized by dividing over the larger of the pair. The largest percentage drop in the series below 'low'

pcaElbow(varpc, low = 0.08, max.pc = 0.9)

Arguments

varpc
numeric, vector of eigenvalues, or 'percentage of variance' explained datapoints for each principle component. If only using a partial set of components, should first pass to estimate.eig.vpcs() to estimate any missing eigenvalues.

low
numeric, between zero and one, the threshold to define that a principle component does not explain much 'of the variance'.

max.pc
maximum percentage of the variance to capture before the elbow (cumulative sum to PC 'n')

Value

The number of last principle component to keep, prior to the determined elbow cutoff

Author

Nicholas Cooper

Examples

# correlated data
mat <- sim.cor(100,50)

#> Error in sim.cor(100, 50): could not find function "sim.cor"

result <- princomp(mat)

#> Error in princomp(mat): object 'mat' not found

eig <- result$sdev^2

#> Error in eval(expr, envir, enclos): object 'result' not found

elb.a <- quick.elbow(eig)

#> Error in quick.elbow(eig): could not find function "quick.elbow"

pca.scree.plot(eig,elbow=elb.a,M=mat)

#> Error in pca.scree.plot(eig, elbow = elb.a, M = mat): could not find function "pca.scree.plot"

elb.b <- quick.elbow(eig,low=.05) # decrease 'low' to select more components

#> Error in quick.elbow(eig, low = 0.05): could not find function "quick.elbow"

pca.scree.plot(eig,elbow=elb.b,M=mat)

#> Error in pca.scree.plot(eig, elbow = elb.b, M = mat): could not find function "pca.scree.plot"

# random (largely independent) data, usually higher elbow #
mat2 <- generate.test.matrix(5,3)

#> Error in generate.test.matrix(5, 3): could not find function "generate.test.matrix"

result2 <- princomp(mat2)

#> Error in princomp(mat2): object 'mat2' not found

eig2 <- result2$sdev^2

#> Error in eval(expr, envir, enclos): object 'result2' not found

elb2 <- quick.elbow(result2$sdev^2)

#> Error in quick.elbow(result2$sdev^2): could not find function "quick.elbow"

pca.scree.plot(eig2,elbow=elb2,M=mat2)

#> Error in pca.scree.plot(eig2, elbow = elb2, M = mat2): could not find function "pca.scree.plot"