Wrapper to run GO enrichment, using fisher enrichment method

Wrapper to run GO enrichment, using fisher enrichment method. Builds on TopGo functionality.

runEnrichment(
  gene.list,
  gene.universe,
  species,
  p.threshold = 0.01,
  p.adj.threshold = 0.05
)

Arguments

gene.list
named list of gene sets to query, where name specify name of gene set, and entry is vector of gene symbols.

gene.universe
background genes

species
Species: Hs or Mm

p.threshold
numeric specifying p value threshold. Default is 0.01.

p.adj.threshold
numeric specifing adjusted p value threshold. Defulat is 0.05.

Value

list of data.frames. results.table.p contains unadjusted results, results.table.bh contains BH-adjusted results.

See also

Author

Nicholas Mikolajewicz

Examples

enrichment.list <- runEnrichment(module.list, gene.universe = all.genes, species = "Hs")

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: 'BiocGenerics'

#> The following objects are masked from 'package:parallel':
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from 'package:base':
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which, which.max, which.min

#> Loading required package: graph

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: GO.db

#> Loading required package: AnnotationDbi

#> Loading required package: stats4

#> Loading required package: IRanges

#> Loading required package: S4Vectors

#> 
#> Attaching package: 'S4Vectors'

#> The following objects are masked from 'package:data.table':
#> 
#>     first, second

#> The following object is masked from 'package:base':
#> 
#>     expand.grid

#> 
#> Attaching package: 'IRanges'

#> The following object is masked from 'package:data.table':
#> 
#>     shift

#> The following object is masked from 'package:grDevices':
#> 
#>     windows

#> Loading required package: SparseM

#> Warning: package 'SparseM' was built under R version 4.0.4

#> 
#> Attaching package: 'SparseM'

#> The following object is masked from 'package:base':
#> 
#>     backsolve

#> 
#> groupGOTerms: 	GOBPTerm, GOMFTerm, GOCCTerm environments built.

#> 
#> Attaching package: 'topGO'

#> The following object is masked from 'package:IRanges':
#> 
#>     members

#> Error in runEnrichment(module.list, gene.universe = all.genes, species = "Hs"): object 'module.list' not found

results.p <- enrichment.list$results.table.p

#> Error in eval(expr, envir, enclos): object 'enrichment.list' not found

results.bh <- enrichment.list$results.table.bh

#> Error in eval(expr, envir, enclos): object 'enrichment.list' not found