Run gene-set enrichment analysis (GSEA)

Run gene-set enrichment analysis (GSEA)

runGSEA(
  gene,
  value,
  species,
  db = "GO",
  my.entrez = NULL,
  my.pathway = NULL,
  min.size = 3,
  max.size = 300,
  do.plot = T,
  plot.top.n = 10,
  path.name.wrap.width = 40
)

Arguments

gene
Character vector of gene names

value
Numeric vector of values used to rank genes (e.g., logFC, expression, etc.). Must be same length as gene argument.

species
Species. One of "Mm" (mouse) or "Hs" (human)

my.entrez
Data.frame with SYMBOL and ENTREZID columns. Used to relate gene symbols to Entrez IDs. Retrieved internally unless explicitly provided.

my.pathway
Named list of pathways, with each entry containing vector of Entrez IDs. Retrieved internally unless explicitly provided.

min.size
Minimum gene set size. Default is 3.

max.size
Minimum gene set size. Default is 300.

do.plot
Logical to return dotplot visualizing top GSEA ranked pathways. Default is T.

plot.top.n
Numeric specifying how many top pathways to visualize. Default is 10.

path.name.wrap.width
Numeric specifying width of pathway names to wrap around. Argument passed to stringr::str_wrap(..., width = path.name.wrap.width)

Value

list of enrichment results

Author

Nicholas Mikolajewicz