Run hypergeometric gene enrichment analysis.

Run hypergeometric gene enrichment analysis.

runHG(
  gene.list,
  gene.universe,
  species,
  pathway.db = "Bader",
  n.workers = 16,
  my.pathway = NULL,
  my.pathway.representation = "ENTREZ",
  min.size = 2,
  max.size = 300,
  e2s = F,
  go.ontology = "BP",
  verbose = T
)

Arguments

gene.list
Named list of genesets to run enrichment on (symbol format).

gene.universe
Gene universe (symbol format)

species
Species. One of "Mm" (mouse) or "Hs" (human)

pathway.db
Name of databse to get annotation genelists from. Default is "Bader". See scMiko::getAnnotationPathways() for options.

n.workers
Number of workers for parallelization. Default is 16.

my.pathway
Named list of pathways, with each entry containing vector of Entrez IDs. Retrieved internally unless explicitly provided.

my.pathway.representation
If my.pathway is provided, species which format, options = "SYMBOL" or "ENTREZ". ENTREZ is Default.

min.size
min geneset size. Default is 2.

max.size
max geneset size. Default is 300.

e2s
entrez to symbol mapping (computationally demanding). Default False.

go.ontology
BP, MF or CC. Ignored if pathway.db != "GO".

verbose
Print progress. Default is TRUE.

Value

enrichment results

Author

Nicholas Mikolajewicz