Run Robust Prinicipal Component Analysis

Run a robust PCA (rPCA) dimensionality reduction on single-cell seurat object.

runRPCA(
  object,
  assay = NULL,
  features = NULL,
  npcs = 50,
  maxpcs = 50,
  reduction.key = "RPC_",
  reduction.name = "rpca",
  seed.use = 42,
  verbose = T,
  method = c("hubert", "robpca", "fasthcs", "pcal"),
  maxdir = 100,
  signflip = T,
  ...
)

Arguments

object
Seurat object

assay
Name of Assay rPCA is being run on

features
Features to compute PCA on. If features=NULL, PCA will be run using scaled features for the Assay. Note that the features must be present in the scaled data. Any requested features that are not scaled or have 0 variance will be dropped, and the PCA will be run using the remaining features.

npcs
Total Number of PCs to compute and store (50 by default)

maxpcs
Max Number of PCs to compute and store (50 by default)

reduction.name
dimensional reduction name, rpca by default

seed.use
Set a random seed. By default, sets the seed to 42. Setting NULL will not set a seed.

verbose
Print progress. Default is TRUE.

method
Robust PCA method. default is "hubert".

maxdir
maximal number of random directions to use for computing the outlyingness of the data points. Default is maxdir=100.

signflip
a logical value indicating wheather to try to solve the sign indeterminancy of the loadings - ad hoc approach setting the maximum element in a singular vector to be positive. Default is signflip = TRUE

...
additional parameters passed to rPCA methods.

reductions.key
dimensional reduction key, specifies the string before the number for the dimension names. RPC by default

Value

Seurat object

See also

Author

Nicholas Mikolajewicz

Examples