Evaluate signature coherence.

Evaluates signature coherence by computing signature score (using runMS) and then calculating correlations between signature components (i.e., gene expression) and signature score. Genes for which correlations do not exceed coherence threshold are dropped and remaining genes are used to calculate new coherent signature.

signatureCoherence(
  object = NULL,
  ms.result = NULL,
  genelist,
  slot = "data",
  assay = DefaultAssay(object),
  coherence.threshold = 0.1,
  show.grid = T,
  coherent.ms = T,
  cor.method = "spearman",
  ...
)

Arguments

object
Seurat object

ms.result
Output from runMS() for provided genelist. If not specified, signatures are computed using runMS.

genelist
Named list of genes used to compute gene signature.

slot
Seurat slot used for gene expression correlations. Default is 'data.'

assay
Seurat assay. Default is DefaultAssay(object).

coherence.threshold
Coherence threshold. Default is 0.1.

show.grid
Logical to show grid on coherence graph.

coherent.ms
Logical to recompute coherent signature following coherence thresholding.

cor.method
correlation metod used to determine signature coherence. Options: spearman, pearson. Default is spearman.

...
additional parameters passed to scMiko::runMS().

Value

list of results

See also

Author

Nicholas Mikolajewicz

Examples