Convert sparse matrix to dense matrix

Convert sparse matrix to dense matrix. Developed to handle large datasets, by constructing dense matrix block-by-block.

sparse2dense(mat.sparse, block.size = 10000, transpose = F, verbose.error = F)

Arguments

mat.sparse

Sparse matrix.

block.size

If large dataset, construct dense matrix column-wise block-by-block, where block size is specified by numeric block.size parameter. Default is 10000.

transpose

Logical specifying whether to transpose data.

verbose.error

Logical specifying whether to print error message in case of large dataset which requires block-by-block construction.

Value

dense matrix

See also

Author

Nicholas Mikolajewicz

Examples

# Get sparse matrix
exp.mat <- so@assays[[current.assay]]@data
#> Error in eval(expr, envir, enclos): object 'so' not found

# convert to dense matrix
mat.dense <- sparse2dense(mat.sparse, transpose = T)
#> Error in sparse2dense(mat.sparse, transpose = T): could not find function "sparse2dense"