Summarize SSN, ICA or NMF gene program/module analyses

Summarize SSN, ICA or NMF gene program/module analyses. Expression heatmaps and gene program enrichments are generated to facilitate analysis and interpretation of gene programs

summarizeModules(
  cell.object,
  gene.object,
  module.type = c("ssn", "ica", "nmf"),
  gene.list = NULL,
  reduction = NULL,
  connectivity_plot = NULL,
  group_by = "seurat_clusters",
  raster = F,
  show.n.pathways = 10,
  n.workers = 1,
  raster.threshold = 10000,
  verbose = T
)

Arguments

cell.object
Seurat object (gene x cell). Same as input to runSSN(object = cell.object, ...).

gene.object
Seurat object (cell x gene) obtained from SSN analysis.

module.type
Which gene program results to summarize

"ssn" - scale-free shared nearest neighbor network

"ica" - independent component analysis

"nmf" - non-negative matrix factorization

gene.list
feature to visualize expression.

reduction
specify which reduction slot use for summary. Ignored if `module.type` = "ssn".

group_by
which meta feature to group by. Default is "seurat_clusters".

raster
whether UMAPs are rasterized (recommended for large datasets).

show.n.pathways
number of pathway annotations to show in each enrichment plot. Default is 10.

n.workers
number of workers for parallel implementation of pathway enrichment analysis. Default is 1.

raster.threshold
cell count threshold at which to switch to rasterized plots. Default is 10000.

verbose
Print progress. Default is T.

connecitivity.plot
"plot_edge" generated by SSNConnectivity function. If not specified, not edges are plotted in SSN graph plot.

Value

List of summarize gene program results, including expression heatmaps and pathway enrichments

References

https://nmikolajewicz.github.io/scMiko/articles/Module_Detection.html

See also

Author

Nicholas Mikolajewicz

Examples