Update central log

Update central log. Central log keeps track of every scPipeline run and includes information about the module run, data used, user, date.

updateCentralLog(
  Module,
  input.data = NA,
  input.subset = NA,
  clog.file = "moduleLog.csv",
  log.path = if (exists("data.path")) data.path else "",
  user.id = if (exists("user")) user else "guest",
  run.notes = NA,
  pdf.flag = save.pdf
)

Arguments

Module
Module Name/ID. e.g., "M01" is used to specify Module 1.

input.data
Input dataset (that was preprocessed in current module). A character.

input.subset
data subset. A character specifying whether a certain subset of data was analyzed. A character.

clog.file
Central log file name. A character.

log.path
Path to central log file. Set to data.path specified in .RProfile by default.

user.id
User name. Set to user specified in .RProfile by default.

run.notes
Additional notes to describe current run.

pdf.flag
Logical indicating whether PDF figures were generated.

Author

Nicholas Mikolajewicz

Examples

# update central log (usually at the end of a successful module run)
updateCentralLog(Module = "M01", pdf.flag = F)

#> Warning: Updating central log...

#> Error in paste0(data.path, clog.file): object 'data.path' not found