Run WNN Multi-Modal Integration.

Run WNN Multi-Modal Integration. Modified wrapper for seurat WNN workflow.

wnn_Run(
  object,
  wnn.knn = 20,
  umap.knn = 20,
  umap.min.dist = 0.1,
  do.scale = F,
  do.center = F,
  normalize.margin = NA,
  pca.thres = 0.9,
  cluster.resolution = 1,
  cluster.algorithm = 3,
  min.pct = 0.25,
  split.var = "seurat_clusters",
  neighborhood.membership = T,
  dist.metric = "euclidean",
  pca.nDim = 50,
  pca.weight.by.var = T,
  ...
)

Arguments

object
Seurat object or list of expression matrices. If seurat object, expression matrices are extracted. If list, assumes that expression matrix entries have column-wise genes and row-wise cells.

wnn.knn
the number of multimodal neighbors to compute. 20 by default

umap.knn
This determines the number of neighboring points used in local approximations of manifold structure. Larger values will result in more global structure being preserved at the loss of detailed local structure. In general this parameter should often be in the range 5 to 50. Default: 20

umap.min.dist
This controls how tightly the embedding is allowed compress points together. Larger values ensure embedded points are moreevenly distributed, while smaller values allow the algorithm to optimise more accurately with regard to local structure. Sensible values are in the range 0.001 to 0.5. Default: 0.1

do.scale
Logical to scale expression. Default is F.

do.center
Logical to center expression. Default is F.

normalize.margin
If specified, normalize across rows/cells (1) or columns/genes (2)

pca.thres
Variance explained threshold for PC component inclusion. Default is 0.9.

cluster.resolution
Cluster resolution for integrated network. Default is 1.

cluster.algorithm
Algorithm for modularity optimization (1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm). See Seurat:FindClusters() for details. Default: 3.

min.pct
Minimum expression fraction for inclusion in network integration. Default is 0.25. Ignored if object is list.

split.var
Grouping variable for expression fraction filter. Default is 'seurat_clusters'. Ignored if object is list.

neighborhood.membership
Logical whether to return list of local neighborhoods. Default: T.

dist.metric
Distance metric for annoy. Options include: euclidean, cosine, manhattan, and hamming

pca.nDim
Number of principal components to consider initially. Default is 50.

pca.weight.by.var
Weight the cell embeddings by the variance of each PC. Default is T.

...
additional parameters passed to Seurat::FindMultiModalNeighbors()

Value

list of integrated results

Author

Nicholas Mikolajewicz