List of scMiko gene sets

Reference gene sets consolidated from numerous sources to faciliate cell annotation and module scoring.

data(geneSets)

Format

A list of data.frames containing the following genesets:

CancerSEA_Hs (link)

Cancer single-cell state atlas

CellMarker_Hs_Zhang2019 (link)

Cell marker atlas: Manually-curated celltype-specific gene sets from over 100,000 published papers. (Zhang et al. (2019) Nucleic Acids Res.)

coreCTL

Core cytotoxic T lymphocyte kill genes. (Lawson et al. (2020) in revision.)

cyclingGenes_Hs_Seurat (link)

Cell-cycle markers (Tirosh et al. (2016) Science)

GBM_Hs_Neftel2019 (link)

Glioblastoma-subtype specific markers (Neftel et al. (2019) Cell)

Immune_Hs_Nirmal2018 (link)

Immune cell signatures for 7 immune cell types identified in tumors using scRNAseq (Nirmal et el. (2018) Cancer Immunol Res.)

MCA_Mm_top50 (link)

Top 50 differentially-markers for each cell in Mouse organogenesis cell atlas characterized using scRNAseq. (Cao et al. (2019) Nature)

MCA_Mm_top100 (link)

Top 100 differentially-markers for each cell in Mouse organogenesis cell atlas characterized using scRNAseq. (Cao et al. (2019) Nature)

neuralDif_Mm_HH

Neural differential markers provided by Hong Han.

Panglao_Mm (link)

Murine cell-type specific markers consolidated from multiple studies through unified framework. (Franzen et al. (2019) Database (Oxford))

Panglao_Hs (link)

Human cell-type specific markers consolidated from multiple studies through unified framework. (Franzen et al. (2019) Database (Oxford))

Renal_Mm_PMID29622724 (link)

Murine kidney cell-specific markers (Park et al. (2018) Science)

TRRUSTv2_Mm (link)

Murine-specific trancription factor genesets; TRRUSTv2 database. (Han et al. (2018) Nucleic Acid Res.)

TRRUSTv2_Hs (link)

Human-specific trancription factor genesets; TRRUSTv2 database. (Han et al. (2018) Nucleic Acid Res.)

TRRUSTv2_regulation_Mm (link)

Murine-specific trancription factor genesets, with regulatory annotation (i.e., activating vs repressing); TRRUSTv2 database. (Han et al. (2018) Nucleic Acid Res.)

TRRUSTv2_regulation_Hs (link)

Human-specific trancription factor genesets, with regulatory annotation (i.e., activating vs repressing); TRRUSTv2 database. (Han et al. (2018) Nucleic Acid Res.)

VastDB_Mm (link)

Bulk RNAseq-derived cell-specific markers (Tapial et al. (2017) Genome Res.)

immuneSubtypes (link)

Tumor immunesubtype markers (Thorsson et al. (2018) Immunity)

immune_literaturePanel

Immune marker panels obtained from literature review

universal_literaturePanel

universal marker panels obtained from literature review

cancer_literaturePanel

cancer marker panels obtained from literature review

AnimalTFDB (link)

Animal transcription factor database. List of human and murine transcription factors

Richards_NatureCancer_2021_sc (link)

Glioblastoma stem cell injury response and developmental genesets. Single cell derived.

Richards_NatureCancer_2021_bulk (link)

Glioblastoma stem cell injury response and developmental genesets. Bulk RNA derived.

Cahoy_JNeurosci

Neural cell gene panels

Liddelow_Nature_2017

Astrocyte gene panels

Zhong_Nature_2018

Neural cell gene panels

Verhaak_CancerCell_2010

Verhaak GBM-subtype genesets

Examples

data(geneSets)
Verhaak_CancerCell_2010 <- geneSets[["Verhaak_CancerCell_2010"]]